[TEST] Remove SessionManagerSuite to reduce duplicated tests
