+++ kyuubi-common/src/test/scala/org/apache/kyuubi/operation/HiveJDBCTestHelper.scala
@@ -19,17 +19,12 @@ package org.apache.kyuubi.operation
 
 import java.sql.ResultSet
 
-import scala.collection.JavaConverters._
-
 import org.apache.hive.service.rpc.thrift._
 import org.apache.hive.service.rpc.thrift.TCLIService.Iface
 import org.apache.hive.service.rpc.thrift.TOperationState._
-import org.apache.thrift.protocol.TBinaryProtocol
-import org.apache.thrift.transport.TSocket
 import org.scalatest.time.SpanSugar.convertIntToGrainOfTime
 
 import org.apache.kyuubi.Utils
-import org.apache.kyuubi.service.authentication.PlainSASLHelper
 
 trait HiveJDBCTestHelper extends JDBCTestHelper {
 
@@ -81,43 +76,13 @@ trait HiveJDBCTestHelper extends JDBCTestHelper {
     jdbcUrl + sessionConfStr + jdbcConfStr + jdbcVarsStr
   }
 
-  def withThriftClient(f: TCLIService.Iface => Unit): Unit = {
-    val hostAndPort = jdbcUrl.stripPrefix("jdbc:hive2://").split("/;").head.split(":")
-    val host = hostAndPort.head
-    val port = hostAndPort(1).toInt
-    val socket = new TSocket(host, port)
-    val transport = PlainSASLHelper.getPlainTransport(Utils.currentUser, password, socket)
-
-    val protocol = new TBinaryProtocol(transport)
-    val client = new TCLIService.Client(protocol)
-    transport.open()
-    try {
-      f(client)
-    } finally {
-      socket.close()
-    }
+  def withThriftClient[T](f: TCLIService.Iface => T): T = {
+    TClientTestUtils.withThriftClient(jdbcUrl.stripPrefix("jdbc:hive2://").split("/;").head)(f)
   }
 
-  def withSessionHandle(f: (TCLIService.Iface, TSessionHandle) => Unit): Unit = {
-    withThriftClient { client =>
-      val req = new TOpenSessionReq()
-      req.setUsername(user)
-      req.setPassword(password)
-      req.setConfiguration(_sessionConfigs.asJava)
-      val resp = client.OpenSession(req)
-      val handle = resp.getSessionHandle
-
-      try {
-        f(client, handle)
-      } finally {
-        val tCloseSessionReq = new TCloseSessionReq(handle)
-        try {
-          client.CloseSession(tCloseSessionReq)
-        } catch {
-          case e: Exception => error(s"Failed to close $handle", e)
-        }
-      }
-    }
+  def withSessionHandle[T](f: (TCLIService.Iface, TSessionHandle) => T): T = {
+    val hostAndPort = jdbcUrl.stripPrefix("jdbc:hive2://").split("/;").head
+    TClientTestUtils.withSessionHandle(hostAndPort, sessionConfigs)(f)
   }
 
   def checkGetSchemas(rs: ResultSet, dbNames: Seq[String], catalogName: String = ""): Unit = {
